feat user data generation for test users

+++ rr/views/testuser.py
@@ -1,4 +1,4 @@
-from rr.models.serviceprovider import ServiceProvider
+from rr.models.serviceprovider import ServiceProvider, SPAttribute
 from rr.models.testuser import TestUser, TestUserData
 from rr.forms.testuser import TestUserForm, TestUserDataForm
 from rr.models.attribute import Attribute
@@ -6,6 +6,41 @@ from django.shortcuts import render
 from django.http.response import Http404
 from django.contrib.auth.decorators import login_required
 from django.utils import timezone
+from random import randint
+
+
+def generate_user_data(testuser, userdata=True, otherdata=True):
+    attributes = SPAttribute.objects.filter(sp=testuser.sp, end_at=None)
+    for attribute in attributes:
+        if userdata:
+            if attribute.attribute.friendlyname == "cn":
+                TestUserData.objects.update_or_create(user=testuser, attribute=attribute.attribute,
+                                                      defaults={'value': testuser.firstname + " " + testuser.lastname})
+            elif attribute.attribute.friendlyname == "givenName":
+                TestUserData.objects.update_or_create(user=testuser, attribute=attribute.attribute,
+                                                      defaults={'value': testuser.firstname})
+            elif attribute.attribute.friendlyname == "sn":
+                TestUserData.objects.update_or_create(user=testuser, attribute=attribute.attribute,
+                                                      defaults={'value': testuser.lastname})
+            elif attribute.attribute.friendlyname == "displayName":
+                TestUserData.objects.update_or_create(user=testuser, attribute=attribute.attribute,
+                                                      defaults={'value': testuser.firstname + " " + testuser.lastname})
+            elif attribute.attribute.friendlyname == "cn":
+                TestUserData.objects.update_or_create(user=testuser, attribute=attribute.attribute,
+                                                      defaults={'value': testuser.firstname + " " + testuser.lastname})
+            elif attribute.attribute.friendlyname == "uid":
+                TestUserData.objects.update_or_create(user=testuser, attribute=attribute.attribute,
+                                                      defaults={'value': testuser.username})
+            elif attribute.attribute.friendlyname == "mail":
+                TestUserData.objects.update_or_create(user=testuser, attribute=attribute.attribute,
+                                                      defaults={'value': testuser.firstname.lower() + "." + testuser.lastname.lower() + "@example.org"})
+            elif attribute.attribute.friendlyname == "eduPersonPrincipalName":
+                TestUserData.objects.update_or_create(user=testuser, attribute=attribute.attribute,
+                                                      defaults={'value': testuser.username + "@example.org"})
+        if otherdata:
+            if attribute.attribute.friendlyname == "funetEduPersonStudentID":
+                TestUserData.objects.update_or_create(user=testuser, attribute=attribute.attribute,
+                                                      defaults={'value': str(randint(100000, 999999))})
 
 
 @login_required
@@ -48,8 +83,11 @@ def testuser_list(request, pk):
                 password = form.cleaned_data['password']
                 firstname = form.cleaned_data['firstname']
                 lastname = form.cleaned_data['lastname']
-                TestUser.objects.create(sp=sp, username=username, password=password, firstname=firstname, lastname=lastname, end_at=None)
+                userdata = form.cleaned_data['userdata']
+                otherdata = form.cleaned_data['otherdata']
+                testuser = TestUser.objects.create(sp=sp, username=username, password=password, firstname=firstname, lastname=lastname, end_at=None)
                 form = TestUserForm(sp=sp)
+                generate_user_data(testuser, userdata, otherdata)
         elif "remove_testuser" in request.POST:
             for key, value in request.POST.dict().items():
                 if value == "on":
@@ -68,7 +106,7 @@ def testuser_attribute_data(request, pk):
     """
     Displays a form including :model:`rr.Attribute` that
     are linked to :model:`rr.ServiceProvider` and values
-    for 
+    for
 
     If value is given, creates :model:`rr.TestUserData`
 
@@ -96,24 +134,30 @@ def testuser_attribute_data(request, pk):
     if not request.user.is_superuser and not ServiceProvider.objects.get(pk=testuser.pk, admins=request.user, end_at=None):
         raise Http404(_("User provided does not exist"))
     if request.method == "POST":
-        form = TestUserDataForm(request.POST, user=testuser)
-        if form.is_valid():
-            for field in form:
-                data = form.cleaned_data.get(field.name)
-                userdata = TestUserData.objects.filter(user=testuser, attribute__friendlyname=field.name).first()
-                if userdata and not data:
-                    userdata.delete()
-                elif data:
-                    if not userdata:
-                        attribute = Attribute.objects.filter(friendlyname=field.name).first()
-                        TestUserData.objects.create(user=testuser,
-                                                    attribute=attribute,
-                                                    value=data)
-
-                    else:
-                        if userdata.value != data:
-                            userdata.value = data
-                            userdata.save()
+        if "save_data" in request.POST:
+            form = TestUserDataForm(request.POST, user=testuser)
+            if form.is_valid():
+                for field in form:
+                    data = form.cleaned_data.get(field.name)
+                    userdata = TestUserData.objects.filter(user=testuser, attribute__friendlyname=field.name).first()
+                    if userdata and not data:
+                        userdata.delete()
+                    elif data:
+                        if not userdata:
+                            attribute = Attribute.objects.filter(friendlyname=field.name).first()
+                            TestUserData.objects.create(user=testuser,
+                                                        attribute=attribute,
+                                                        value=data)
+                        else:
+                            if userdata.value != data:
+                                userdata.value = data
+                                userdata.save()
+        if "reset_userdata" in request.POST:
+            generate_user_data(testuser, userdata=True, otherdata=False)
+            form = TestUserDataForm(user=testuser)
+        if "reset_otherdata" in request.POST:
+            generate_user_data(testuser, userdata=False, otherdata=True)
+            form = TestUserDataForm(user=testuser)
     else:
         form = TestUserDataForm(user=testuser)
     return render(request, "rr/testuser_attribute_data.html", {'form': form,
